fix: purge 시 문제 참조 S3 키 삭제 방지 #60

+++ src/main/java/cmc/delta/domain/problem/adapter/in/worker/ScanPurgeWorker.java
@@ -124,6 +124,9 @@ public class ScanPurgeWorker extends AbstractClaimingScanWorker {
 		}
 		List<cmc.delta.domain.problem.model.asset.Asset> assets = assetJpaRepository.findAllByScan_Id(scanId);
 		for (cmc.delta.domain.problem.model.asset.Asset asset : assets) {
+			if (problemRepository.existsByOriginalStorageKey(asset.getStorageKey())) {
+				continue;
+			}
 			try {
 				storagePort.deleteImage(asset.getStorageKey());
 			} catch (Exception e) {
